Assistant checkpoint: Fix database schema and timestamp handling
Assistant generated file changes:
- migrations/0000_init.sql: Fix timestamp data type for PostgreSQL compatibility
- shared/schema.ts: Update schema to use bigint for timestamps
- server/storage.ts: Fix timestamp handling in storage

---

User prompt:

We have since added a database. Are we still deployment ready?

+++ migrations/0000_init.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS "user_sessions" (
 	"session_id" text NOT NULL,
 	"form_data" json,
 	"results" json,
-	"created_at" integer NOT NULL,
+	"created_at" bigint NOT NULL,
 	CONSTRAINT "user_sessions_session_id_unique" UNIQUE("session_id")
 );
